MDEV-21469: Implement crash-safe binary logging of the user XA
Make XA PREPARE, XA COMMIT and XA ROLLBACK statements
crash-safe when --log-bin is specified.

At execution of XA PREPARE, XA COMMIT and XA ROLLBACK their replication
events are firstly written into the binary log and then submitted to
storage engines.
In a case server crashes after writing to binary log but before all
involved engines have processed the command, the following recovery
will execute the command's replication events. In case of
multiple-engine XA PREPARE for instance that would first involve
a total rollback and the total replay afterward.

On the implementation level the recovery time binary log parsing
is augemented to pay attention to
the user XA xids to identify the XA transactions' state:s in binary log, and
eventually match them against their states in engines, see
MYSQL_BIN_LOG::recover_explicit_xa_prepare().
In discrepancy cases the outdated state in the engines is corrected with
resubmitting the transaction prepare group of events, or completion
ones. The multi-engine XA also conducts rollback in prepared engines
prior to re-submission.
The fact of multipe-engine involved is registed into
Gtid_log_event::flags2 as one bit. The boolean value is
precise to deal with two engines in an XA transaction. With more
than 2 recoverable engines the flag method is still correct though
may be pessimistic, as it treats all recoverable engines as XA
participants. So when the number of Engines greater
than the number of prepared engines for an XA that XA is treated
as partially completed.
There's no new bit allocated in flags2, an existing one is reused,
observing that XA "COMPLETE" does not require multi-engine hint
for its recovery.